+++ brew/src/main/java/ch/fhnw/brew/business/service/CustomerService.java
@@ -2,8 +2,10 @@ package ch.fhnw.brew.business.service;
 
 import ch.fhnw.brew.data.domain.Customer;
 import ch.fhnw.brew.data.repository.CustomerRepository;
+import ch.fhnw.brew.data.repository.OrderRepository;
 import ch.fhnw.brew.exception.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,6 +15,9 @@ public class CustomerService {
 
     @Autowired
     private CustomerRepository customerRepository;
+
+    @Autowired
+    private OrderRepository orderRepository; 
 
     public Customer addCustomer(Customer customer) {
         return customerRepository.save(customer);
@@ -35,9 +40,16 @@ public class CustomerService {
         return customerRepository.save(customer);
     }
 
+    @PreAuthorize("hasRole('ADMIN')")
     public void deleteCustomer(Integer id) {
         Customer customer = customerRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Customer does not exist"));
+
+        boolean hasOrders = orderRepository.existsByCustomerCustomerID(id);
+        if (hasOrders) {
+            throw new IllegalStateException("Customer cannot be deleted because they have existing orders");
+        }
+
         customerRepository.delete(customer);
     }
