+++ brew/src/main/java/ch/fhnw/brew/business/service/RecipeService.java
@@ -1,9 +1,11 @@
 package ch.fhnw.brew.business.service;
 
 import ch.fhnw.brew.data.domain.Recipe;
+import ch.fhnw.brew.data.repository.BrewingProtocolRepository;
 import ch.fhnw.brew.data.repository.RecipeRepository;
 import ch.fhnw.brew.exception.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,6 +15,9 @@ public class RecipeService {
 
     @Autowired
     private RecipeRepository recipeRepository;
+
+    @Autowired
+    private BrewingProtocolRepository brewingProtocolRepository;
 
     public Recipe addRecipe(Recipe recipe) {
         return recipeRepository.save(recipe);
@@ -28,9 +33,15 @@ public class RecipeService {
         return recipeRepository.save(recipe);
     }
 
+    @PreAuthorize("hasRole('ADMIN')")
     public void deleteRecipe(Integer id) {
         Recipe recipe = recipeRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Recipe not found"));
+
+        if (brewingProtocolRepository.existsByRecipe(recipe)) {
+            throw new IllegalStateException("Cannot delete recipe: it is still used by brewing protocols.");
+        }
+
         recipeRepository.delete(recipe);
     }
